Moved tests in a different file and changed them
